requirements.txt
flask==3.1.2
gunicorn==23.0.0
ultralytics==8.3.190
opencv-python-headless==4.12.0.88
numpy==2.0.2
tencentcloud-sdk-python==3.0.937
